@@ -3,10 +3,10 @@ apply plugin: 'com.android.library'
 dependencies {
     testImplementation 'junit:junit:4.12'
 
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
 
     // rxjava
-    implementation 'io.reactivex.rxjava2:rxjava:2.1.4'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.0'
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
     implementation 'com.jakewharton.rxrelay2:rxrelay:2.0.0'
 }
@@ -16,7 +16,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3'
     }
 }
@@ -37,7 +37,7 @@ ext {
     siteUrl = 'https://github.com/novo-dimaporo/archx'
     gitUrl = 'https://github.com/novo-dimaporo/archx.git'
 
-    libraryVersion = '0.0.4'
+    libraryVersion = '0.0.6'
 
     developerId = 'novo-dimaporo'
     developerName = 'Novo Dimaporo'
@@ -50,8 +50,8 @@ ext {
 
 ext.versions = [
         'minSdk'        : 14,
-        'compileSdk'    : 27,
-        'targetSdk'     : 27,
+        'compileSdk'    : 28,
+        'targetSdk'     : 28,
         'buildTools'    : '27.0.3',
 
         'release'       : libraryVersion,
